Research paper
The research paper stemming from this project was presented at the distinguished International Student Conference on Multidisciplinary & Current Technical Research hosted by the Madhav Institute of Technology and Science (MITS), Gwalior. This platform provided an opportunity to emphasize the pivotal role of employing machine learning techniques in the early detection of lung cancer.